Define marks only once
Instead of defining all marks explicitly, now use a loop construction
instead. This requires a mapping between tiptap marks and Font Awesome
names, hence the object.

In order to keep TypeScript happy, for the toggle method we must cast as
a `Capitalize<T>`, so that is confident the function really exists. To
get this, we must first extract the keys of the object. TypeScript then
also automatically checks if the function exists, preventing bugs.