Create custom Tiptap extensions
Instead of using the ones provided by default, started developing own
extensions that behave similarly. The main reason to do this, is so that
we're able to create one uniform API for commands. That is: if we create
a command bar, it's easy to trigger specific commands for each
extension, as well as detecting if a certain mode is active or commands
should become active.

Everything is quite straightforward, except for `list` and `list-item`.
These required direct usage of ProseMirror, which is the editor Tiptap
wraps.

The extensions aren't complete yet, but this was a good enough
foundation to commit and push. With this, enough confidence was built
that this route will work in the long run.